Task 189: split burndown calculations into separate classes
