+++ src/main/java/taiga/util/burndown/BurndownCalculator.java
@@ -1,0 +1,9 @@
+package taiga.util.burndown;
+
+import taiga.model.query.sprint.Sprint;
+
+import java.util.List;
+
+public interface BurndownCalculator {
+    List<BurnDownEntry> calculate(Sprint sprint);
+}
